# src/main/resources/api.yml
openapi: "3.0.3"
info:
  version: 1.0.0
  title: SDC Service API
  license:
    name: Apache License 2.0
servers:
  #- url: http://34.159.3.234:8080/
  - url: http://localhost:8080/
paths:
  /valuation/legacy/value:
    post:
      summary: Request mapping for the value
      operationId: legacyValue
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LegacyValueRequest"
      responses:
        "200":
          description: Value was calculated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValueResult"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /valuation/legacy/margin:
    post:
      summary: Request mapping for the margin
      operationId: legacyMargin
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LegacyMarginRequest"
      responses:
        "200":
          description: Margin was calculated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarginResult"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /valuation/legacy/test/product:
    post:
      summary: Request mapping for the value of a product (using fixed market data)
      operationId: testProductValue
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - tradeData
              properties:
                tradeData:
                  type: string
                  format: binary
      responses:
        "200":
          description: Value was calculated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValueResult"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
    ValueResult:
      $ref: "schemas/openapi-schemas/ValueResult.yml"
    MarginResult:
      $ref: "schemas/openapi-schemas/MarginResult.yml"
    Error:
      $ref: "schemas/openapi-schemas/Error.yml"
...
    LegacyValueRequest:
      $ref: "schemas/openapi-schemas/LegacyValueRequest.yml"
    LegacyMarginRequest:
      $ref: "schemas/openapi-schemas/LegacyMarginRequest.yml"
...